requirements.txt
git+https://github.com/iDevision/enhanced-discord.py@925539d3eb4913492bdb4afd12776d8679fa645e
python-dotenv
nest_asyncio
